Bugfix in the bubble sort algorithm
The bubble sort algorithm was not a true bubble sort. It was going into
recursion before the largest value had bubbled to the end because it was
calling bubble sort every time any value was not greater than the next
value. It will now correctly bubble each largest value to the end.